Refactor: move reusable stuff to utils
